requirements.txt
aiodns==3.0.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
discord.py==2.3.1
frozenlist==1.3.3
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-texttospeech==2.14.1
googleapis-common-protos==1.59.1
grpcio==1.56.0
grpcio-status==1.56.0
idna==3.4
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
multidict==6.0.4
proto-plus==1.22.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycares==4.3.0
pycparser==2.21
PyNaCl==1.5.0
referencing==0.30.0
requests==2.31.0
rpds-py==0.9.2
rsa==4.9
six==1.16.0
urllib3==1.26.16
yarl==1.9.2
